+++ src/frontend/src/expr/type_inference/mod.rs
@@ -17,11 +17,15 @@
 use std::collections::HashMap;
 use std::vec;
 
-use itertools::{iproduct, Itertools};
+use itertools::iproduct;
 use risingwave_common::error::{ErrorCode, Result};
 use risingwave_common::types::DataType;
 
-use crate::expr::{Expr as _, ExprImpl, ExprType};
+use crate::expr::ExprType;
+
+mod cast;
+mod func;
+pub use cast::{align_types, cast_ok, least_restrictive, CastContext};
 
 /// `DataTypeName` is designed for type derivation here. In other scenarios,
 /// use `DataType` instead.
@@ -370,146 +374,6 @@ lazy_static::lazy_static! {
     };
 }
 
-/// Find the least restrictive type. Used by `VALUES`, `CASE`, `UNION`, etc.
-/// It is a simplified version of the rule used in
-/// [PG](https://www.postgresql.org/docs/current/typeconv-union-case.html).
-///
-/// If you also need to cast them to this type, and there are more than 2 exprs, check out
-/// [`align_types`].
-pub fn least_restrictive(lhs: DataType, rhs: DataType) -> Result<DataType> {
-    if lhs == rhs {
-        Ok(lhs)
-    } else if cast_ok(&lhs, &rhs, &CastContext::Implicit) {
-        Ok(rhs)
-    } else if cast_ok(&rhs, &lhs, &CastContext::Implicit) {
-        Ok(lhs)
-    } else {
-        Err(ErrorCode::BindError(format!("types {:?} and {:?} cannot be matched", lhs, rhs)).into())
-    }
-}
-
-/// Find the `least_restrictive` type over a list of `exprs`, and add implicit cast when necessary.
-/// Used by `VALUES`, `CASE`, `UNION`, etc. See [PG](https://www.postgresql.org/docs/current/typeconv-union-case.html).
-pub fn align_types<'a>(exprs: impl Iterator<Item = &'a mut ExprImpl>) -> Result<DataType> {
-    use std::mem::swap;
-
-    let exprs = exprs.collect_vec();
-    // Essentially a filter_map followed by a try_reduce, which is unstable.
-    let mut ret_type = None;
-    for e in &exprs {
-        if e.is_null() {
-            continue;
-        }
-        ret_type = match ret_type {
-            None => Some(e.return_type()),
-            Some(t) => Some(least_restrictive(t, e.return_type())?),
-        };
-    }
-    let ret_type = ret_type.unwrap_or(DataType::Varchar);
-    for e in exprs {
-        let mut dummy = ExprImpl::literal_bool(false);
-        swap(&mut dummy, e);
-        *e = dummy.cast_implicit(ret_type.clone())?;
-    }
-    Ok(ret_type)
-}
-
-/// The context a cast operation is invoked in. An implicit cast operation is allowed in a context
-/// that allows explicit casts, but not vice versa. See details in
-/// [PG](https://www.postgresql.org/docs/current/catalog-pg-cast.html).
-#[derive(Debug, Eq, Ord, PartialEq, PartialOrd)]
-pub enum CastContext {
-    Implicit,
-    Assign,
-    Explicit,
-}
-
-/// Checks whether casting from `source` to `target` is ok in `allows` context.
-pub fn cast_ok(source: &DataType, target: &DataType, allows: &CastContext) -> bool {
-    let k = (name_of(source), name_of(target));
-    matches!(CAST_MAP.get(&k), Some(context) if context <= allows)
-}
-
-fn build_cast_map() -> HashMap<(DataTypeName, DataTypeName), CastContext> {
-    use DataTypeName as T;
-
-    // Implicit cast operations in PG are organized in 3 sequences, with the reverse direction being
-    // assign cast operations.
-    // https://github.com/postgres/postgres/blob/e0064f0ff6dfada2695330c6bc1945fa7ae813be/src/include/catalog/pg_cast.dat#L18-L20
-    let mut m = HashMap::new();
-    insert_cast_seq(
-        &mut m,
-        &[
-            T::Int16,
-            T::Int32,
-            T::Int64,
-            T::Decimal,
-            T::Float32,
-            T::Float64,
-        ],
-    );
-    insert_cast_seq(&mut m, &[T::Date, T::Timestamp, T::Timestampz]);
-    insert_cast_seq(&mut m, &[T::Time, T::Interval]);
-
-    // Casting to and from string type.
-    for t in [
-        T::Boolean,
-        T::Int16,
-        T::Int32,
-        T::Int64,
-        T::Decimal,
-        T::Float32,
-        T::Float64,
-        T::Date,
-        T::Timestamp,
-        T::Timestampz,
-        T::Time,
-        T::Interval,
-    ] {
-        m.insert((t, T::Varchar), CastContext::Assign);
-        // Casting from string is explicit-only in PG.
-        // But as we bind string literals to `varchar` rather than `unknown`, allowing them in
-        //  assign context enables this shorter statement:
-        // `insert into t values ('2022-01-01')`
-        // If it was explicit:
-        // `insert into t values ('2022-01-01'::date)`
-        // `insert into t values (date '2022-01-01')`
-        m.insert((T::Varchar, t), CastContext::Assign);
-    }
-
-    // Misc casts allowed by PG that are neither in implicit cast sequences nor from/to string.
-    m.insert((T::Timestamp, T::Time), CastContext::Assign);
-    m.insert((T::Timestampz, T::Time), CastContext::Assign);
-    m.insert((T::Boolean, T::Int32), CastContext::Explicit);
-    m.insert((T::Int32, T::Boolean), CastContext::Explicit);
-    m
-}
-
-fn insert_cast_seq(
-    m: &mut HashMap<(DataTypeName, DataTypeName), CastContext>,
-    types: &[DataTypeName],
-) {
-    for (source_index, source_type) in types.iter().enumerate() {
-        for (target_index, target_type) in types.iter().enumerate() {
-            let cast_context = match source_index.cmp(&target_index) {
-                std::cmp::Ordering::Less => CastContext::Implicit,
-                // Unnecessary cast between the same type should have been removed.
-                // Note that sizing cast between `NUMERIC(18, 3)` and `NUMERIC(20, 4)` or between
-                // `int` and `int not null` may still be necessary. But we do not have such types
-                // yet.
-                std::cmp::Ordering::Equal => continue,
-                std::cmp::Ordering::Greater => CastContext::Assign,
-            };
-            m.insert((*source_type, *target_type), cast_context);
-        }
-    }
-}
-
-lazy_static::lazy_static! {
-    static ref CAST_MAP: HashMap<(DataTypeName, DataTypeName), CastContext> = {
-        build_cast_map()
-    };
-}
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -647,102 +511,5 @@ mod tests {
         for (expr, num_t) in iproduct!(exprs, num_types) {
             test_infer_type_not_exist(expr, vec![num_t, DataType::Boolean]);
         }
-    }
-
-    fn gen_cast_table(allows: CastContext) -> Vec<String> {
-        use itertools::Itertools as _;
-        use DataType as T;
-        let all_types = &[
-            T::Boolean,
-            T::Int16,
-            T::Int32,
-            T::Int64,
-            T::Decimal,
-            T::Float32,
-            T::Float64,
-            T::Varchar,
-            T::Date,
-            T::Timestamp,
-            T::Timestampz,
-            T::Time,
-            T::Interval,
-        ];
-        all_types
-            .iter()
-            .map(|source| {
-                all_types
-                    .iter()
-                    .map(|target| match cast_ok(source, target, &allows) {
-                        false => ' ',
-                        true => 'T',
-                    })
-                    .collect::<String>()
-            })
-            .collect_vec()
-    }
-
-    #[test]
-    fn test_cast_ok() {
-        // With the help of a script we can obtain the 3 expected cast tables from PG. They are
-        // slightly modified on same-type cast and from-string cast for reasons explained above in
-        // `build_cast_map`.
-
-        let actual = gen_cast_table(CastContext::Implicit);
-        assert_eq!(
-            actual,
-            vec![
-                "             ", // bool
-                "  TTTTT      ",
-                "   TTTT      ",
-                "    TTT      ",
-                "     TT      ",
-                "      T      ",
-                "             ",
-                "             ", // varchar
-                "         TT  ",
-                "          T  ",
-                "             ",
-                "            T",
-                "             ",
-            ]
-        );
-        let actual = gen_cast_table(CastContext::Assign);
-        assert_eq!(
-            actual,
-            vec![
-                "       T     ", // bool
-                "  TTTTTT     ",
-                " T TTTTT     ",
-                " TT TTTT     ",
-                " TTT TTT     ",
-                " TTTT TT     ",
-                " TTTTT T     ",
-                "TTTTTTT TTTTT", // varchar
-                "       T TT  ",
-                "       TT TT ",
-                "       TTT T ",
-                "       T    T",
-                "       T   T ",
-            ]
-        );
-        let actual = gen_cast_table(CastContext::Explicit);
-        assert_eq!(
-            actual,
-            vec![
-                "  T    T     ", // bool
-                "  TTTTTT     ",
-                "TT TTTTT     ",
-                " TT TTTT     ",
-                " TTT TTT     ",
-                " TTTT TT     ",
-                " TTTTT T     ",
-                "TTTTTTT TTTTT", // varchar
-                "       T TT  ",
-                "       TT TT ",
-                "       TTT T ",
-                "       T    T",
-                "       T   T ",
-            ]
-        );
     }
 }
